extract buildDeployment to separate file
